fix GetInt function optimize ease of use add back arrow update dependencies, TypeScript

+++ internal/app/handlers/functions.go
@@ -31,8 +31,8 @@ func GetStringArray(request map[string]interface{}, name string) []string {
 // GetInt get int or 0
 func GetInt(request map[string]interface{}, name string) int {
 	// cast and return value
-	value, _ := request[name].(int)
-	return value
+	value, _ := request[name].(float64)
+	return int(value)
 }
 
 // GetBool get string or false
